Adding skipif for valgrind
This future started passing for PrgEnv-cray on Brad's triage watch.
It's getting lucky, as a valgrind run on the "correct" binary shows
the use of uninitialized data.

I've added a skipif to only run the test when we are using valgrind
(knowing that currently this means we won't ever test it).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21123 3a8e244f-b0f2-452b-bcba-4c88e055c3ca