Added a bug future on sync variables deleted prematurely when passed by ref and used in 'begin' blocks.
Currently, the following idiom is implemented correctly (albeit memory-leaky)
by not deleting/freeing x$ at the end of something():

  proc something() {
    var x$: sync int;
    begin {
      ... x$ ...
    }
  }

However, if x$ is passed as an argument and declared in the caller,
it will be deleted/freed in the caller:

  proc something(x$: sync int) {
    begin {
      ... x$ ...
    }
  }

  proc caller() {
    var x$: sync int;
    something(x$);
    // compiler adds delete x$;
  }

This is incorrect because x$ in the 'begin' may reference
already-deleted memory.

The only reliable way to observe the bug at run time is by running
these tests under valgrind. The tests contain a barrier to ensure the
sync variables are referenced after they are deallocated.

This is actually a "suite" of several tests, all variants on the above theme:

* autodestroy-with-begin-1.chpl

It is is the above base case, renamed, with the barrier and an
intermediate function/call between caller() and something() added.

* autodestroy-with-begin-ref.chpl

It is a variation on autodestroy-with-begin-1.chpl that specifies
the 'ref' intent explicitly for the sync variable formals.
This behavior should be identical to blank intent, but currently
it triggers an unimplemented case in the compiler.

* autodestroy-with-begin-2.chpl

It is a variation on autodestroy-with-begin-1.chpl that adds another
sync variable used in an uninteresting way, for comparison.

* autodestroy-with-begin-3.chpl

It is a variation on autodestroy-with-begin-1.chpl that has a lot more
intermediate functions/calls, declares additional sync variables along the
way, and has an additional 'begin' that's invoked from the original 'begin'.
This is to ensure that the fix to this bug works when the declaration
of a sync variable and the begin where it's used are far apart.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21126 3a8e244f-b0f2-452b-bcba-4c88e055c3ca