Removed all 'make depend' rules to opt for more modern equivalents
For developers (CHPL_DEVELOPER=true), the most important things to
know about this commit are:

a) You should 'make clean all' after updating to it

b) You should never need to use 'make depend' again

c) If I've done things right, you should never need to use 'make
   clean' or 'make clobber' as part of your normal development cycle.

In more detail:

This commit removes the annoying/slow distinct 'make depend' step that
has been used by Chapel developers to compute dependences on demand.
Instead, it relies on gcc-oriented capabilities to compute dependences
during compilation.  This has the benefit of (a) reducing overhead by
computing dependences as we go and (b) removing complicated ordering
requirements that were present in the original system due to, e.g.,
relying on third-party packages to be installed in the right place
before dependences could be computed correctly.

The downside of this patch is that for compilers that do not have a
gcc-like capability, we are removing the ability to do a 'make depend'
step.  The reason this isn't a huge concern is that most Chapel
developers use gcc (to the extent that many compilers
have never had correct 'make depend' capabilities enabled), and that
most users don't use 'make depend' (and it seems several developers do
not either :).  Over time, we could re-instate a separate 'make
depend' step that would use some other capability to compute
dependences, simply by having it use the same .d filename convention
that we're switching to here.  But for clarity/simplicity, this commit
simply removes that capability for the time being to clean the slate a
bit.

At present, I've only added this capability when CHPL_*_COMPILER is
set to gnu.  It can easily be set for other compilers by defining
the DEPEND_CFLAGS variable in make/compiler/Makefile.*

In the new system, when DEPEND=1 during the building of the compiler
or runtime (which it is by default when CHPL_DEVELOPER is set), -MMD
and -MP are added to the compiler command line which causes the
non-system #include dependences of the source file foo.c to be written
to the same directory as the foo.o file was in the file foo.d.  We
then '-include' the *.d files in that directory at the bottom of the
relevant Makefiles.

The -MMD flag does most of the above.  The -MP flag adds a phony
dependence for the header files themselves, e.g., it will say
something like:

          foo.h:

The benefit of this is that if foo.h goes away or is renamed as the
result of a change or commit, the Makefiles won't complain that they
don't know how to make foo.h.  Of course, the next compile will update
the .d file and remove the dependence on foo.h in such a case.

The other way to clean up stale dependences like this is to run the
'make cleandeps' step which will clean all the .d files out of the
current object file directories.  And of course, make clobber will
remove the directories themselves as always so that's another way
to clean up.

The constructive steps to enable this (and files affected) were:

* Added DEPEND_CFLAGS for the gnu compiler:

       make/compiler/Makefile.gnu

* Added 'cleandep' rules where they weren't already defined

       Makefile

* Added logic to add DEPEND_CFLAGS to CFLAGS when DEPEND is 1:

       make/Makefile.base

* Added logic to turn DEPEND to 1 when CHPL_DEVELOPER is true

       compiler/make/Makefile.compiler.head
       runtime/make/Makefile.runtime.head

* Updated SVN_DEPENDS definition to refer to *.d files rather than
  Makefile.svn.depend

       compiler/make/Makefile.compiler.head
       runtime/make/Makefile.runtime.head

* For some reason with this change, the .timestamp file that's used to
  help create the object file directories safely/sanely was getting
  removed as though it was a more implicit rule than it used to be.
  Running with --debug I couldnt' find any rationale for the new
  behavior, so I added .SECONDARY: clauses to preserve these files and
  avoid the noise of re-making/re-removing them at each step.

       make/Makefile.base
       runtime/make/Makefile.runtime.foot

* Updated documentation and notes on Makefiles; also moved Makefile.NOTES
  into doc/developer which is a better place for it (it didn't exist back
  in the dark ages when this was originally written).

       compiler/make/Makefile.NOTES
       doc/release/README.building
       doc/developer/bestPractices/CHPL_DEVELOPER.txt

* Updated a very out-of-date comment

       make/Makefile.base


The destructive steps required to remove and clean up 'make depend'
were:

* Hollowed out 'depend' targets and made the top-level one print a
  message to help those of us who type 'depend' habitually break
  the habit; also removed supporting variables/rules related to
  supporting 'make depend'

       Makefile
       compiler/make/Makefile.compiler.head
       compiler/make/Makefile.compiler.rules
       compiler/make/Makefile.compiler.subdirrules
       compiler/make/Makefile.compiler.foot
       compiler/parser/Makefile
       runtime/Makefile
       runtime/make/Makefile.runtime.head
       runtime/make/Makefile.runtime.subdirrules
       runtime/src/qio/regexp/re2/Makefile
       runtime/src/qio/regexp/re2/Makefile.share
       runtime/src/comm/gasnet/Makefile.share
       runtime/src/mem/tcmalloc/Makefile.share
       runtime/make/Makefile.runtime.foot
       runtime/make/Makefile.runtime.emptydirrules
       runtime/make/Makefile.runtime.rules

* Removed the old MAKEDEPEND/CMAKEDEPEND rules from all the compilers

       make/compiler/Makefile.pathscale
       make/compiler/Makefile.llvm-gcc
       make/compiler/Makefile.ibm
       make/compiler/Makefile.cray-vec
       make/compiler/Makefile.intel
       make/compiler/Makefile.cray-mta
       make/compiler/Makefile.pgi
       make/compiler/Makefile.tile-cc
       make/compiler/Makefile.clang
       make/compiler/Makefile.gnu
       make/compiler/Makefile.tau
       make/compiler/Makefile.cray-x-series




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21130 3a8e244f-b0f2-452b-bcba-4c88e055c3ca