Reinstate locale->localeID->locale short-circuit optimization.
Looks for places where the expression supplied as the "on" expression x is of
type dtLocale.  In these cases, x can be passed directly to the wrapon_fn,
rather than going through
    x = chpl_localeID_to_locale(__primitive("_wide_get_locale", x))
first.

This version is implemented only for the blocking side.
It needs to be expanded to the nonblocking side as well.


Details:

    runtime/include/chpl-gen-includes.h
    modules/internal/ChapelLocale.chpl
Added chpl_gen_localeID_to_int() to support localeID comparisons.
 
    modules/internal/RootLocale.chpl
Do not perform locale lookup if we are on the right locale (by ID) already.

    compiler/AST/primitive.cpp
_is_here returns a Boolean.

    compiler/passes/parallel.cpp
Implemented the optimization, using findDestLocale().
Removed unused collapseLocaleLookup code.  This was a prototype that is now
absorbed in the find... functions.
Pulled def-use map out of findLocaleLookup() so we can use it in findDestLocale() too.


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21136 3a8e244f-b0f2-452b-bcba-4c88e055c3ca