Added locale-lookup short-circuit optimization to nonblocking on clauses.
The optimization proceeds in four phases:

Important parts of the on call are identified using a def-use map on the block
containing it.  This is done up-front because after the locale operand is
removed from the direct call to on_fn in the left branch of the local on
conditional, def-use maps cannot be computed again until after bundleArgs() is
run.  That problem is due to the resulting formal/actual argument mismatch.
bundleArgs() trims out formal locale argument from on_fn, so the correct
formal/actual match is restored.

If the on expression was originally a locale expression, looking up its locale
given its locale ID is pointless, so that lookup is bridged out.  This is now
done whether the call is blocking or not.

For blocking calls, the earlier optimization of calling the on_fn in the
running task is performed next (blocking-local optimization).

The locale-lookup short-circuit optimization is completed by removing the unused
locale lookup call.  Otherwise, it is moved into the right branch of the
blocking-local conditional.  For the nonblocking case, it is removed from the
tree and then put right back where it was.

    compiler/passes/parallel.cpp
Renamed optimizeLocalOns as passArgsToNestedFns, since that's what it really does.
Pulled out the optimization stuff as optimizeOnCall().
Now we match with something of a type derived from dtLocale, not just objects
with dtLocale as their declared type.
Added a lot of comments.
 


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21140 3a8e244f-b0f2-452b-bcba-4c88e055c3ca