Capturing tests related to bugzilla entry #794133
Bugzilla entry #794133 (under Cray's system) captures a case
in which we're improperly freeing a domain prior to its end
of life due to reference counting it based on the number of
elements having that domain.  In particular, given:

  var A: [1..3] [1..2] real;

we currently reference count the anonymous domain {1..2}
3 times, once per index in the outer domain {1..3}.  The
trouble comes when the outer domain is empty, for example:

  var D = {1..0};
  var A: [D] [1..2] real;

In this case, the domain {1..2} gets reference counted zero times and
is therefore freed when it leaves scope even though a later
re-assignment to D may require it to be alive.  The solution here
relates to keeping {1..2} alive as long as that type is alive, not
(simply) the number of elements relying on it.

I've written a simplification of this test (simple.chpl) and a number
of variations on it that exercise various corner cases and am
capturing them here.  I need to do further testing under trunk to
determine which are futures and which are not, so have marked the
directory NOTEST for now.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21146 3a8e244f-b0f2-452b-bcba-4c88e055c3ca