Clean up build functions for array and domain literals
[reviewed by diten]

Changed the name of chpl__build*ArrayLiteral to chpl__build*ArrayExpr
to match chpl__buildDomainExpr.  We had some discussion in chat about
naming, which resulted in choosing Expr over Literal.  We weren't
completely happy with Expr, but decided to go for consistency first.

Also, I somehow missed the removal of chpl__buildArrayLiteralWarn in
the confusion of splitting r21004 and r21005 into 2 commits, so remove
it this time for sure.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21155 3a8e244f-b0f2-452b-bcba-4c88e055c3ca