Hardcoding fix to avoid re-autoconf-ing qthreads
[reviewed by Dylan Stark]

With the most recent version of Qthreads, we were running into
troubles where building it would cause it to want to re-autoconf
itself, but our autoconf tools weren't of a recent enough version
for that to work as intended.

The root cause of this was that by checking Qthreads into svn
and then back out again, the time stamps of the various files
were scrambled up, so the dependences would indicate that
certain autoconf-related steps needed to be re-run.  Our
hypothesis is that this probably occurred on previous Qthreads
releases as well, but that we didn't have the version
incomptability problem, so weren't noticing it.

In this commit, we hard-code some rules into our Makefile,
similar to what GASNet does with its Bootstrap script.
It touches some key files in a specific order to make sure
that autoconf isn't re-run.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21164 3a8e244f-b0f2-452b-bcba-4c88e055c3ca