Enabled back bulk transfer on arrays of ints, reals, imaginaries, which got inadvertently disabled in r21165.
This fixes the regression in optimizations/sungeun/bulk_transfer/assign.chpl

The issue is that the previous set of chpl__supportedDataTypeForBulkTransfer
functions was inadvertently returning 'false' on the above types - because:

  proc chpl__supportedDataTypeForBulkTransfer(x: complex) param return false;
  proc chpl__supportedDataTypeForBulkTransfer(x) param return true;

When invoked on anything that has an implicit type conversion to complex
(i.e. ints, reals, imaginaries), the 'false' case is more specific,
even though it was unintended.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21167 3a8e244f-b0f2-452b-bcba-4c88e055c3ca