Fix capacityParSafe test to complete more reliably. [reviewed by diten]
This test fails occasionally in the nightly tests.  It sometimes runs out of
resources (memory or available threads), which is not the main point of the
test.  The main point is to create contention on the capacity element in an
associative domain descriptor by trying to update it under parallel execution.

The change involves replacing a 1000-wide coforall containing one update per
task with a 100-wide coforall containing 100 updates per task.  This presumably
increases the chances for contention by a factor of 10 without exceeding the
practical limit of a few hundred parallel tasks.

At the same time, the requested capacity was reduced to (i+j), to keep
the total amount of allocated memory within reasonable bounds.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21168 3a8e244f-b0f2-452b-bcba-4c88e055c3ca