Removing tuple()
[reviewed by diten]

With the introduction of 1-tuple syntax in the last release (x,)
the tuple() function becomes much less useful than it used to be,
so this commit removes it from standard use.  Of course, a user
who particularly likes having such a function can define it
themselves, and we could define it as an optional library
capability if we chose to.  I'm not convinced it's useful enough
to be necessary, though.

In particular, if you used to write:

   tuple(x)     -- you should now write --> (x,)
   tuple(x,y,z) -- you should now write --> (x,y,z)

This commit removes the tuple() function and performs the
rewrites above for all tests and internal modules.  It also
updates the spec and spec tests w.r.t. 1-tuples.

Along the way, I renamed a few variables named 'tuple'
which were not very descriptive, though arguably my new
names aren't a ton better.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21172 3a8e244f-b0f2-452b-bcba-4c88e055c3ca