Added a bug future where compiler crashes with an improper AST when the program defines two recursive leader iterators.
This bug is being fixed by Tom.

Credits: Tom discovered and diagnosed the bug, explained it in writing and
additionally to me directly. Steve wrote the test that I am using to
describe the bug. Tom picked that test for illustration. I use some
of Tom's writing in the .future.

From .future:

The issue is that the lowerIterators pass creates these two functions,
when there is a recursive iterator:

  chpl__copyRecursiveIteratorArgumentBundle
  chpl__freeRecursiveIteratorArgumentBundle

whose _loopBodyFnArgs argument contains a reference to the iterator
argument bundle.

Only a single copy of each of these is created. The type of the
_loopBodyFnArgs argument is that for the first iterator that causes
these methods to be created.

So when these methods are invoked for the second iterator that needs them,
it passes its own argument bundle, whose type does not match that for
the first iterator, resulting in a violation of an AST invariant,
which is cheerfully reported by the compiler as:

  error: actual formal type mismatch [expr.cpp:2587]

Interestingly, the bodies of these methods are already created
to handle multiple iterator bundle types, so it is only the header
that is an issue.

The .future also contains some background details for the above.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21189 3a8e244f-b0f2-452b-bcba-4c88e055c3ca