Fix to array-of-arrays regression
[reviewed by sungeun]

My array-of-array fix (r21192) ran into problems for the test
test/arrays/deitz/part4/test_array_of_associative_arrays.chpl
due to a reference-after-free problem that only showed up in
valgrind testing and certain parallel settings.  This commit
fixes that in a way that could leak slightly more memory in
some cases, but which would be hard to fix until Sung's ref
counting commits are made.

The issue is that at the end of the program, we:

1) autodestroy defaultDist.  Its refcount goes to 0 so it's
   freed.

2) then we autodestroy D and A.  The free of A causes my new
   decrement reference counts code to get called which creates
   a new temporary array which refers to defaultDist and tries
   to decrement its reference count even though it is (a) 0
   and (b) already freed.

We discussed a few ways of improving this situation, including:

a) autodestroying variables in the opposite order than they are
   allocated (this wouldn't actually help the problem, but seems
   more intuitive and could remove the need for any reference
   counting in certain well-structured situations in the future)

b) determining when defaultDist's reference count is getting bumped
   up and down.  The linksDistribution is a domain-centric way of
   doing this, but the question of whether or not a reference
   count is bumped seems like it should really be a property of
   the distribution, not the domain (at least, according to Sung's
   and my understanding and Steve's original commit message).  Sung
   is going to look into refactoring this as part of her commit
   where she's also wrestling with freeing distributions.

c) either remove the decrement of a distribution's reference count in
   ~_distribution or guard it by the replacement for
   linksDistribution(), or find a corresponding place to increment it.
   It seems that it's getting decremented without getting incremented
   when my new inc/dec routines are creating their temporary arrays.

In the end, it seemed like the best way to proceed was to bump the
reference count of the domain's distribution by one in my new
...incRefCounts.. routine in order to make sure that it's kept
alive as long as the domain is.  This didn't seem to affect the
memory used in the example directory and should only cause leaks
in the case that arrays-of-arrays are created in which the array
elements have a distribution other than defaultDist (and part of
our argument was that we're already leaking a bunch of memory
there anyway).

Put another way, for the time being, correctness seemed better than
being leak-free at this point, particularly since all the code
relating to plugging leaks for domains, arrays, and distributions is
being rewritten currently anyway.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21199 3a8e244f-b0f2-452b-bcba-4c88e055c3ca