Do no invalidate the dtObject type even if it si unused after function resolution.
We simplify the compiler code slightly by removing the special treatment of
dtObject in this regard.  That's because we do not see much gain in invalidating
dtObject even if it is unused, and also we expect it to always be used as the
parent of 'locale'.

We ensure the dtObject is always valid by adding and respecting the
FLAG_GLOBAL_TYPE_SYMBOL flag.  In that case, the assignments "dtObject = NULL;"
and their surrounding conditionals are moot.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21201 3a8e244f-b0f2-452b-bcba-4c88e055c3ca