Fix to address bug 794131 (flag extern array arguments as errors)
[reviewed by Vass]

This commit addresses bug 794131 in which the use of an array formal
type in an extern proc declaration (something we don't support yet)
triggered an internal error in some cases and a C compilation error in
others. As a result of this commit, it now gets flagged as an error
and refers the user to README.extern.

As we've discussed, ultimately it would be nice if a formal array type
in an extern was interpreted as "a C array" and that passing a Chapel
array to that argument would cause the Chapel array to remove all its
wrapping and pass the dense buffer of data out to C, if that makes
sense for the given domain map (e.g., for DefaultRectangular or a
1-locale Block it might; for a multi-locale block, it arguably would
not).  This commit should be considered a placeholder to mark this
case as unsupported until such time as we support that feature.

The approach I took was pretty straightforward (though a little
regrettable): In building the function during parsing, if the function
is extern, we scan the formal arguments for array types, as indicated
by calls to chpl__buildArrayRuntimeType().  The regrettable bit is
that I used a strcmp() to determine this; but I didn't see a similarly
straightforward way to cleanly catch these cases short of refactoring
all places that build such a call into a single routine and applying a
flag. That would be a nice change, but felt like a lot to change a day
before the code freeze.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21205 3a8e244f-b0f2-452b-bcba-4c88e055c3ca