Reverted a part of r21049 by changing the warning "accessing the locale of a non-lvalue expression" back to "accessing the locale of a local expression".
Brad explains that the warning's intention was to warn the user about
doing something provably meaningless, such as 'on here {...}'.
In which case the pre-r21049 message is more appropriate.

Notes:

* The warning is currently not completely implemented.  For example,
the test test_dot_locale_error.chpl does not print all the warnings
we'd like to have. This is independent of r21049 or this commit.

* Almost each (or simply each?) non-lvalue expression has an associated
locale, e.g. the locale where that 'const' is declared or where that literal
is computed. So technically it is possible to query a locale of an l-value.

* The user may intentionally query the locale of a non-lvalue expression
or of a local expression. We intend to provide a way for users
to suppress these warnings when desired.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21207 3a8e244f-b0f2-452b-bcba-4c88e055c3ca