Add reserved words for new symbols in PGI C runtime library version 13.3.
New symbols showed up in the C runtime library delivered with the latest
release of the PGI compiler.  Some of these conflict with symbols defined in our
Chapel test code, causing standard regression tests to fail.
A selection of these symbols have been added to the list of reserved symbol
names, so the code generator can avert likely conflicts.

Only a subset of the symbols published by the libpgc library were added.  These
are the selection critera:
 - Symbols starting with "_", "__", "pgi_", "omp_" and "smp_" were ignored.
 - Symbols containing 7 or more characters were ignored.

The list that remained contains 22 symbols.  None of these are duplicates with
those already in the reservedSymbolNames file.  (One existing duplicate, "int",
was found and removed.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21210 3a8e244f-b0f2-452b-bcba-4c88e055c3ca