Fixed the recursive iterator inlining bug. Un-futurized the corresponding tests.
Most of the work in this commit was done by Tom.
Writing was done jointly by Tom and Vass.
Some of the text comes from recursive-iterator-twice-local.future.


---------------------------------
WHAT WAS THE BUG?

This bug exhibited itself as a BaseAST::verify() assertion failure:
formal/actual argument type mismatch.

The bug involved use of the shared functions:

  chpl__copyRecursiveIteratorArgumentBundle()
  chpl__freeRecursiveIteratorArgumentBundle()

These functions were created with signatures that included the
argument bundle type for the iterator with which they were first
called.  The failure occured when one of these functions was called
with an argument bundle of a different type, for a subsequently-
processed iterator. The bodies of these methods were already generated
to handle multiple iterator bundle types, only the headers were an issue.

Our tests had not exercised this case.  Tom hit it when playing around
with hierarchical architectural descriptions.

As of this commit, the tests under $CHPL_HOME/test that cause the
compiler to create these two functions are only these:

  // these do not exercise the different-argument-bundles case
  functions/deitz/iterators/leader_follower/test_recursive_leader*.chpl
  // these DO exercise that case
  functions/iterators/vass/recursive-iterator-twice-*.chpl


---------------------------------
BACKGROUND

Recursive iterator inlining takes the body of the iterator and turns
this into a function that is installed the global function table.  The
body is then referred to by its function index.  As with other
functions in that table, the arguments are bundled into a class, so
they can be referred to through a single void*.  Code in the caller
bundles up the args, and code in a wrapper function unbundles the args
so they can be passed to the actual body function seriatim.

The global function table was created a while back to replace function
pointers for functions that can be called across locales. See:

  ftableMap/ftableVec
  PRIM_FTABLE_CALL
  r15370
  r16348

To make it more concrete, consider the test test_recursive_leader3.chpl.
Here is an exerpt from the AST after lowerIterators, compiled with
--no-local (without --fast) prior to this commit,
where 'bar' is the name of a parallel iterator:

  type _fn_arg_bundle[1079410] var _arg1[1079436]:_ref(_ir_bar)[743010]

  function chpl__copyRecursiveIteratorArgumentBundle[1079451](arg _loopBodyFnID[1079453]:int(64)[70], arg _loopBodyFnArgs[1079455]:_fn_arg_bundle[1079410]) : _fn_arg_bundle[1079410]
  {
    var _tmp[1079457]:_fn_arg_bundle[1079410]
    .....
    if (_loopBodyFnID == 0)
    {
      ... handle the case where the function index == 0
      (1079494 'move' _tmp[1079457](1079491 'cast' _fn_arg_bundle[1079410] _tmp[1079466]))
    }
    if (_loopBodyFnID == 1)
    {
      ... handle the case where the function index == 1
      (1081603 'move' _tmp[1079457](1081600 'cast' _fn_arg_bundle[1079410] _tmp[1081520]))
    }
    (1079462 return _tmp[1079457])
  }

  function _rec_iter_fn_bar[1079406](arg _ic[1079680]:_ic_bar[743481], arg _loopBodyFnID[1079696]:int(64)[70], arg _loopBodyFnArgs[1079698]:_fn_arg_bundle[1079410]) : void[17]
  {
    .....
          var _tmp[1081470]:int(64)[70]
          var _tmp[1081471]:_fn_arg_bundle[1079410]
          _tmp[1081470] gets _loopBodyFnID
          _tmp[1081471] gets _loopBodyFnArgs
          (1081490 'move' _tmp[1081482](1081486 call chpl__copyRecursiveIteratorArgumentBundle[1079451] _tmp[1081470] _tmp[1081471]))
          (1081495 'move' _tmp[1081484](1081492 'cast' _fn_arg_bundle[1079410] _tmp[1081482]))
    .....
  }

Here, the type of _tmp[1081471] is specific to the iterator in whose
context the call to chpl__copyRecursiveIteratorArgumentBundle() occurs.
So if there are several such iterators, there will be an argument type
mismatch.

The actual type of the argument bundle is only of interest in the
bundling and unbundling code.  In the call to the function table entry
and the signature of the wrapper function (what is actually called),
we can use a more general type (like void*).

The existing code used the actual argument bundle type. This works
fine in other cases because each recursive loop body function has its
own entry in the function table, and both the caller and the wrapper
function know the right type. The two functions in question:

  chpl__copyRecursiveIteratorArgumentBundle()
  chpl__freeRecursiveIteratorArgumentBundle()

which are akin to autoCopy and autoDestroy, are shared among all
iterators. They include a switch on _loopBodyFnID (the index into the
function table) to select the desired clause for the current iterator.


---------------------------------
SOLUTION

The simple solution was to create a type that is used to cast the
actual argument bundles into and out of, to match the actuals and the
formals for the two functions. This type is dtArgBundle or "chpl_argBundle".
The headers for these functions thus become:

  chpl_argBundle chpl__copyRecursiveIteratorArgumentBundle(int loopBodyFnID, chpl_argBundle loopBodyFnArgs)
  void chpl__freeRecursiveIteratorArgumentBundle(int loopBodyFnID, chpl_argBundle loopBodyFnArgs)

The callers cast their bundles into chpl_argBundle. The loopBodyFnArgs
formal is then cast to the appropriate bundle type within each clause
of the switch that's inside each function.


---------------------------------
DISCUSSION

* The ideal solution would provide a separate pair of these functions
for each iterator/loopBodyFnID. That way, no casting and no switching
would be necessary. But this was too much of a change to undertake
at the moment.

* A simpler alternative to the current solution would be to leave the
headers of these two functions as-is, and use the bundle type that it
happened to have - instead of chpl_argBundle - to cast into and out
of. This would be less principled (because of the arbitrary type
choice), but would obviate the need in chpl_argBundle.

* Yet another solution would be to use void* instead of chpl_argBundle.
For some reason that did not flow smoothly through the compiler.

* These two functions make it into the generated code
only when compiling with --fast --no-local. Otherwise
they are removed in the 'prune' pass.

* Not directly related to this fix, but we looked at the flags
FLAG_GLOBAL_TYPE_SYMBOL and FLAG_PRIMITIVE_TYPE...

FLAG_GLOBAL_TYPE_SYMBOL: is used to make a symbol survive pruneResolvedTree(),
where it is consistent with isAlive().

It is inconsistent with prune(), however. Consider this scenario:
 - Say prune() removes something marked with FLAG_GLOBAL_TYPE_SYMBOL
   (which currently does not happen, we think).
 - After the pass comes cleanAST(). It does not clean it up because of
   the flag (see isAlive).
 - But then verify() sees that symbol is not in AST but is not cleaned up.
   It will raise an error.

FLAG_PRIMITIVE_TYPE: is used to make a symbol survive prune().
It is not consulted in cleanAST() or pruneResolvedTree().

TODO: clarify these two flags, make them more consistent.


---------------------------------
DETAILS

    compiler/include/type.h
    compiler/AST/type.cpp
Add the new dtArgBundle type.
Mark it with FLAG_GLOBAL_TYPE_SYMBOL to prevent it from ever being pruned.

    compiler/resolution/lowerIterators.cpp
Use the new class dtArgBundle to represent argument bundles passed to
functions representing the loop body of inlined recursive iterators.  This
allows the actual type of the argument bundles to vary while obeying the AST
invariant that formal and actual types line up, and also producing correct
backend C code.

Rename _loopBodyFnID and _loopBodyFnArgs to remove the leading underscores
(which we should not be using in the generated C code).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21215 3a8e244f-b0f2-452b-bcba-4c88e055c3ca