Moved the test recursive-iterator-twice-multiloc.chpl to test/multilocale and disabled recursive-iterator-twice-local.chpl.
The -multiloc test IMO deserves inclusion in abbreviated multilocale testing
(which, as of the current convention in the Chapel group,
includes test/multilocale but not test/functions).

The -local test is a simplification of the -multiloc test,
so it's probably not worth testing separately.
However, -local may come handy for debugging,
should there be any issues with -multiloc.

To recap:

* Both tests are for the bug described and fixed in r21215.

* This test is simpler, but does not have 'on':

    functions/iterators/vass/recursive-iterator-twice-local.chpl

* This test has 'on', providing more thorough testing:

    multilocale/vass/recursive-iterator-twice-multiloc.chpl


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21216 3a8e244f-b0f2-452b-bcba-4c88e055c3ca